@@ -8,7 +8,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.10"
+python = ">=3.9,<3.10"
 pydantic = "^1.9.1"
 pyproj = "^3.4.0"
 networkx = {extras = ["default"], version = "^2.8.6"}
@@ -19,6 +19,7 @@ pytest = "^7.1.3"
 ipykernel = "^6.13.0"
 black = "^22.3.0"
 isort = "^5.10.1"
+pre-commit = "^2.20.0"
 
 [tool.black]
 skip-string-normalization = true
